refactor: discard generic host implementation
