+++ crates/context/interface/src/lib.rs
@@ -8,6 +8,7 @@ extern crate alloc as std;
 pub mod block;
 pub mod cfg;
 pub mod context;
+pub mod host;
 pub mod journaled_state;
 pub mod local;
 pub mod result;
@@ -18,6 +19,7 @@ pub use cfg::{Cfg, CreateScheme, TransactTo};
 pub use context::{ContextError, ContextSetters, ContextTr};
 pub use database_interface::{DBErrorMarker, Database};
 pub use either;
+pub use host::{DummyHost, Host};
 pub use journaled_state::JournalTr;
 pub use local::{FrameStack, FrameToken, LocalContextTr, OutFrame};
 pub use transaction::{Transaction, TransactionType};
