+++ crates/interpreter/src/lib.rs
@@ -12,8 +12,6 @@ mod macros;
 
 /// Gas calculation utilities and constants.
 pub mod gas;
-/// Host interface for external blockchain state access.
-pub mod host;
 /// Context passed to instruction implementations.
 pub mod instruction_context;
 /// Instruction execution results and success/error types.
@@ -32,8 +30,8 @@ pub use context_interface::{
     context::{SStoreResult, SelfDestructResult, StateLoad},
     CreateScheme,
 };
+pub use context_interface::{host, Host};
 pub use gas::{Gas, InitialAndFloorGas};
-pub use host::Host;
 pub use instruction_context::InstructionContext;
 pub use instruction_result::*;
 pub use instructions::{instruction_table, Instruction, InstructionTable};
